Moved nodes to another file from model
